FIXED – Safari bug #8
Safari bug: pictures are not scrolled but dragged with the mouse #8

This was actually an oversight in the demo (only on the flicker call.)

There is a use case where we allow drag and drop between streams -- so we allow dragging in some cases.  I have disabled for the flicker demo.